+++ src/main/java/org/example/service/PageService.java
@@ -1,4 +1,4 @@
-package org.example;
+package org.example.service;
 
 import java.util.ArrayDeque;
 import java.util.Deque;
